# Significantly speed up glyph lookups with caches