Thanks for this great library, a lot of fun to work with! :)

I found current implementation quite slow when handling fonts with large number of glyphs, for example, u8g2_wqy GB2312 (7539 glyphs).

I benchmarked text drawing time the `Shennong` example on my esp8266 and obtained:
```
Frame #338, avg 566ms, worst 864ms
```

After enabled the `last_unicode` lookup cache, the performance imporoved by 27.5% (1.38x speedup):
```
Frame #338, avg 410ms, worst 615ms
```

This patch significantly improves the performance by 85.9% (7.1x speedup):
```
Frame #338, avg 80ms, worst 105ms
```

Two new caching features are implemented:
1. Fixed point unicode lookup cache
    This feature is responsible for the majority of speedup.
    I divided the glygh space into a configurable N+1 partitions, which significantly reduces linear scan.
    The trade-off is more RAM consumption (6 bytes per partition).
    It is enabled by setting `U8G2_WITH_GLYPH_LOOKUP_CACHE` to 2 (the default), and set a desired number of indexes in `U8G2_INDEX_CNT_UNICODE`.
    - Using index count = 4, consumes 6*4 = 24 bytes, and speed up is 4.3x;
    - The default index count is 8, which consumes 6*8 = 48 bytes, and speed up is 7.1x;
    - Using index count = 16, consumes 6*16 = 96 bytes, and speed up is 12.9x;
    - Using index count = 32, consumes 6*32 = 192 bytes, and speed up is 20.2x;

2. In-memory glyph data cache
    Use memory to buffer glyph data (memory consumption depending on font)
    Currently this cache has negligible improvement on performance (at least on esp8266) and is turned off by default.
    Maybe it can result in speed up under different platform. Also, I plan to implement file based font in the near future, and this cache could significantly speed up (or, prevent major slow-down) in that scenario.
    I put it in this PR just because it is just another aspect of caching, but please let me know if you don't like and I will remove it.