@Adam5Wu 2018-04-28
The last encoding cache mechanism was already there, it was disabled by __unix__ macro, however, when I review how existing code use the __unix__ macro, I found no consistent pattern, so I though it was just an abandoned macro...