@Adam5Wu 2018-04-29
Refactored code to another branch, let me close this one and redo another PR